feat(LocationModal.tsx): add LocationModal component for selecting and accepting a location feat(ArtistServiceProduct.tsx): add ArtistServiceProduct component for displaying and selecting artist services feat(pick-location.tsx): remove LocationModal component from pick-location.tsx and import it from its own file
refactor(pick-more.tsx): remove unused imports and update component names
refactor(pick-datetime.tsx): add state key to setSearchParams for better navigation
refactor(pick-location.tsx): remove unused imports and update component names, add state key to setSearchParams for better navigation

chore(routes): remove unused imports and code in ($locale).treatments.$productHandle.pick-more.tsx
feat(routes): add support for selecting a beauty expert in ($locale).treatments.$productHandle.pick-username.tsx
chore(routes): remove unused imports and code in ($locale).treatments.$productHandle.tsx
