+++ app/routes/($locale).treatments.$productHandle.pick-more.tsx
@@ -1,11 +1,5 @@
-import {Badge, Flex, SimpleGrid, Skeleton, Text} from '@mantine/core';
-import {
-  Await,
-  useLoaderData,
-  useParams,
-  useSearchParams,
-} from '@remix-run/react';
-import {Money, parseGid} from '@shopify/hydrogen';
+import {Flex, SimpleGrid, Skeleton} from '@mantine/core';
+import {Await, useLoaderData, useParams} from '@remix-run/react';
 import {defer, type LoaderFunctionArgs} from '@shopify/remix-oxygen';
 import {Suspense} from 'react';
 import {type ProductItemFragment} from 'storefrontapi.generated';
@@ -13,9 +7,7 @@ import {getBookingShopifyApi} from '~/lib/api/bookingShopifyApi';
 import {type CustomerProductBase} from '~/lib/api/model';
 import {ALL_PRODUCTS_QUERY} from './($locale).artist.$username._index';
 
-import {ArtistServiceCheckboxCard} from '~/components/artist/ArtistServiceCheckboxCard';
-import {TreatmentServiceContent} from '~/components/treatment/TreatmentServiceContent';
-import {durationToTime} from '~/lib/duration';
+import {ArtistServiceProduct} from '~/components/artist/ArtistServiceProduct';
 
 export function shouldRevalidate() {
   return false;
@@ -60,10 +52,6 @@ export default function ArtistTreatments() {
   const {products, services} = useLoaderData<typeof loader>();
   const {productHandle: selectedProductHandle} = useParams();
 
-  const [searchParams] = useSearchParams();
-
-  const shippingId = searchParams.get('shippingId');
-
   return (
     <Suspense
       fallback={
@@ -100,7 +88,11 @@ function RenderArtistProducts({
   const restProductsMarkup = products
     .filter((product) => product.handle !== selectedProductHandle)
     .map((product) => (
-      <ArtistProduct key={product.id} product={product} services={services} />
+      <ArtistServiceProduct
+        key={product.id}
+        product={product}
+        services={services}
+      />
     ));
 
   return (
@@ -111,74 +103,5 @@ function RenderArtistProducts({
         </SimpleGrid>
       </Flex>
     </>
-  );
-}
-
-type ArtistProductProps = {
-  product: ProductItemFragment;
-  services: CustomerProductBase[];
-  defaultChecked?: boolean;
-};
-
-function ArtistProduct({
-  product,
-  services,
-  defaultChecked,
-}: ArtistProductProps) {
-  const [searchParams, setSearchParams] = useSearchParams();
-
-  const onChange = (checked: boolean) => {
-    const newSearchParams = new URLSearchParams(searchParams);
-
-    const existingItems = newSearchParams.getAll('productIds');
-    const itemIndex = existingItems.indexOf(parseGid(product.id).id);
-
-    if (checked && itemIndex === -1) {
-      newSearchParams.append('productIds', parseGid(product.id).id);
-    }
-
-    if (!checked) {
-      const updatedItems = existingItems.filter(
-        (item) => item !== parseGid(product.id).id,
-      );
-      newSearchParams.delete('productIds');
-      updatedItems.forEach((item) =>
-        newSearchParams.append('productIds', item),
-      );
-    }
-
-    setSearchParams(newSearchParams);
-  };
-
-  const artistService = services.find(({productId}) => {
-    return productId.toString() === parseGid(product.id).id;
-  });
-
-  const leftSection = (
-    <Text c="dimmed" size="xs" tt="uppercase" fw={700}>
-      {durationToTime(artistService?.duration ?? 0)}
-    </Text>
-  );
-
-  const rightSection = artistService?.price && (
-    <Badge variant="light" color="gray" size="md">
-      <Money data={artistService?.price as any} />
-    </Badge>
-  );
-
-  return (
-    <ArtistServiceCheckboxCard
-      value={artistService!.productId.toString()}
-      defaultChecked={defaultChecked}
-      onChange={onChange}
-      name="productIds"
-    >
-      <TreatmentServiceContent
-        product={product}
-        description={artistService?.description}
-        leftSection={leftSection}
-        rightSection={rightSection}
-      />
-    </ArtistServiceCheckboxCard>
   );
 }
